requirements.txt
# Core dependencies for division-service
Flask==2.3.3
Flask-CORS==4.0.0
gunicorn==21.2.0

# Base de datos
psycopg2-binary==2.9.7
SQLAlchemy==2.0.21

# GCP SDK
google-cloud-storage==2.10.0
google-cloud-pubsub==2.18.4
google-cloud-workflows==1.14.0
google-cloud-logging==3.8.0
google-auth==2.23.3
google-cloud-core==2.3.3

# Utilidades
python-dotenv==1.0.0
requests==2.31.0
python-dateutil==2.8.2

# Logging estructurado
python-json-logger==2.0.7
structlog==23.1.0

# Security
cryptography==41.0.4

# Validación
marshmallow==3.20.1

# Testing
pytest==7.4.2
pytest-cov==4.1.0

# Development
black==23.9.1
flake8==6.1.0
isort==5.12.0
